Initial proof of concept (script to merge PRs on opted-in repos)
This commit sets up the basic Rubocop and RSpec boilerplate, and
a first pass at the necessary files to implement an auto-merge
service that satisifies [RFC-156](https://github.com/alphagov/govuk-rfcs/blob/main/rfc-156-auto-merge-internal-prs.md).

It introduces a `config/repos_opted_in.yml` file, which will be
the gateway to opting in repos to the auto-merge behaviour. For
now, it contains a single test repo.

Running the `bin/merge_dependabot_prs.rb` script fetches all open
PRs for the repos in that file, and uses the `AUTO_MERGE_TOKEN`
token to authenticate and action merging said PRs using the GitHub
API.

The next steps will be to add the safeguarding checks so that we
only merge PRs that satisfy the strict criteria outlined in RFC
156. We can then consider triggering the automerge script through
GitHub Actions (on a schedule), and opting in production repos.

Trello: https://trello.com/c/RblylctX/3136-build-the-auto-merge-service-for-dependabot-prs-5