Add `all_proposed_updates_semver_allowed?` method
This digs deeper into the proposed changes and reports on whether
or not the proposed dependency version is an allowed one. For
instance, if an app has opted in to 'minor' updates on a given
dependency, but Dependabot is attempting to bump to a 'major' new
version, then this method would return false.

If the app hasn't specified a policy for a particular dependency,
and that dependency appears in the list of proposed changes, then
this method will 'fail open', i.e. the app hasn't specified a
semver limitation, so it will return true for it. Whilst this might
feel more risky, we're already 'failing closed' on the
`all_proposed_dependencies_on_allowlist?` sister method. If we
ever want to allow auto-merge on Dependabot PRs that also bump
subdependencies that may not be on the allowlist (quite a common
occurrence), then we should only have to alter the sister method
to allow that, leaving this one untouched.