Fix error when dependency is added/removed rather than updated
Trying this out on a real PR, a runtime error was raised due to
the assumption that every dependency has both a previous_version
and a next_version. In reality, some Dependabot PRs will add or
remove subdependencies, where there will only be one half of the
equation provided.

Initially I added a nil check to the PullRequest class
`tell_dependency_manager_what_dependabot_is_changing` method.
However, it's really more of a DependencyManager concern. So I've
rearchitected the DependencyManager class to cope with more
procedural input, i.e. we only have to tell it that dependencies
are being removed/added, and leave it to DependencyManager to
figure out whether that constitutes an 'update' or a removal or
addition.

This also leads to cleaner tests in `dependency_manager_spec.rb`,
as many of the tests don't care about versions, so it always felt
odd to have to specify both a previous and next version in those
cases.

Trello: https://trello.com/c/RblylctX/3136-build-the-auto-merge-service-for-dependabot-prs-5