Rename to govuk-dependabot-merger
Previous name: github-action-dependabot-auto-merge.
It was named when we were still considering whether it should be
built as a reusable GitHub Action workflow and triggered externally
from GOV.UK PRs. The 'github-action' prefix would be appropriate
for that scenario.

However, this repo will now run on a schedule. It uses GitHub
Actions to drive that, but could just as easily be invoked from
a different tool, such as Jenkins. The 'github-action' prefix now
exposes an implementation detail that shouldn't really be a part
of the service name.

Removing the prefix leaves us with 'dependabot-auto-merge', but
the 'auto' is implied (we're programmers, after all, seeking to
automate things). 'merge' is a verb, which doesn't describe the
service as well as a noun would ('merger'). Finally, a 'govuk'
prefix helps indicate that this is very much intended for a GOV.UK
engineer audience, as the repo makes several assumptions about the
safety of auto-merging Dependabot PRs based on GOV.UK standards
such as high test coverage (see RFC-156).