Check that all updated dependencies are maintained by GOV.UK
Whilst we stipulate in RFC-156 (and in the README of this repo)
that the Dependabot Merger service should only be enabled for
internal dependencies (e.g. rubocop-govuk), we had no mechanism
in place to prevent developers from auto-merging _external_
dependencies. This commit adds a method to the Dependency Manager
to check that all dependencies being proposed for updating are
in fact owned by the `govuk` user:
https://rubygems.org/profiles/govuk

Note that it should not be possible for malicious actors to grant
'govuk' as owners of their own malicious gems, as it requires
action on our part to accept the invite:
https://guides.rubygems.org/managing-owners-using-ui/#:~:text=Adding%20user%20as%20an%20owner,link%20to%20confirm%20the%20ownership.

Trello: https://trello.com/c/RblylctX/3136-build-the-auto-merge-service-for-dependabot-prs-5