Ensure CI workflow passes
For a long time, GOV.UK had a convention whereby repositories had
to specify a status check called 'test', which should be the name
of the CI job that runs a repo's tests, linter, and so on.
These would all run under the 'test' job, usually via something
like `bundle exec rake`.

In January 2023, a [new convention](https://github.com/alphagov/content-data-admin/commit/89a888bf1d9d303cff50ae65ac9c0821c5c17e93)
was established, whereby there would now be an overall workflow
called 'CI'. Within this workflow, there would be several individual
jobs, each with their own responsibility (e.g. 'test-ruby' for
running tests, 'lint-ruby' for running the linter, and so on).

We should therefore no longer look for a 'test' job: the modern
equivalent is to check for the 'CI' workflow, and to validate
that every job within the CI workflow passes.

Note that as a [stopgap](https://github.com/alphagov/govuk-dependabot-merger/pull/21),
we temporarily stopped looking for a 'test' job and switched to
looking for a 'test-ruby' job instead. This is suboptimal as it
only checks for the outcome of the unit tests, and doesn't
consider other checks such as the linter, so it will incorrectly
[try and fail to approve and merge](https://github.com/alphagov/support-api/pull/870#pullrequestreview-1794218971)
PRs whose unit tests pass but where there is another check failing.