Add "PR doctor" script
This script gives an easy way of populating a GitHubClient instance
with API information for a specific PR, and a way of querying the
PullRequest class for information. This can be a useful way of
debugging why a particular PR did or did not auto-merge as expected
(which otherwise requires a lot of faffing around to get to that
starting point).