Don't crash when encountering unexpected YAML structure
govuk-dependabot-merger was raising a `Psych::SyntaxError` exception
when encountering invalid YAML. See
https://github.com/alphagov/whitehall/blob/9509f9d7c297faf3338535e08c28bc927dfd0b12/.govuk_dependabot_merger.yml,
which has one dependency incorrectly outdented on line 15.

We now handle these errors and skip over the repo.

This also splits out the old `validate_external_config_file`
method into separate methods for validating the presence of, and
contents of, the config file. This was in order to have more
useful log text for teams wanting to find out why their PR was
not automerged.