Introduce class ChangeSet
The purpose of this is to encapsulate the parsing of Dependabot commit
messages and storage of the parsed data (these responsibilities are
currently shared between PullRequest and DependencyManager). This should
let us reduce the coupling between PullRequest (and its spec) and
DependencyManager.

I've also reimplemented the commit parser; Dependabot includes
machine-readable metadata in its commit messages that we can use instead
of parsing the human-readable content & the gemfile. We also don't
really care about the versions of the dependencies, we only care about
the update type (major vs minor vs patch), and Dependabot includes this
in the commit metadata.

I've bumped the dependency on rubocop-govuk to >= 4.16, because this
version disables the `Style/MultilineBlockChain` cop which was falsely
flagging `lib/change_set.rb`.

Finally, I've introduced class Dependency to encapsulate the gem ownership
check.