Rename DependencyManager to PolicyManager
In v2, it will no longer make sense for this class to simply
return an array of absolute dependencies that are allowed to
be auto-merged. V2 is much more flexible than that: repos will
be able to opt into auto-merging all dependencies by default
(whether they're internal or external).

So this class will become more of a 'policy manager' / enforcer,
than just a dependency manager.