Make AutoMerger a module
Like we did to GitHubClient in 1a70f5cbf7c059351d607a0688e6e3fb875f80a0,
there are no fields or instance methods so there'd be no point
in initialising this class.

I believe `merge_dependabot_prs` was originally an instance method
to try to make an engineer think twice about calling the method
outside of `invoke_merge_script!` (which has the bank holiday
protection). But it's still trivial to suffix `.new` to
`AutoMerger`, so this was no protection.