Improve `analyse_dependabot_pr` output and add tests while here
`ruby bin/pr_doctor.rb https://<Dependabot URL>` did not give
great output - it just said whether or not a PR was eligible for
merging, and directed the engineer to add a `byebug` statement to
investigate further. We already have the code to give much more
specific feedback, so let's reuse it in the new `merge_dependabot_pr`
method (with a `dry_run` parameter, because we don't actually want
to approve/merge the PRs we're analysing).