Move `.govuk_dependabot_merger.yml` file checks into AutoMerger
This file doesn't exist in the Dependabot Pull Request, and the
check is identical between all Dependabot PRs in the repo, so it
makes sense to surface this check at a higher level.

The result is that we avoid even making the API call to fetch
the Dependabot PRs, if the remote config file is not valid. This
should really speed up the script!

But the bigger benefit is that we're just one step away from
removing PolicyManager from the PullRequest class altogether.
It makes more sense to invert the dependency, i.e. have the
PolicyManager instance query the PullRequest instance for what it
needs, in the AutoMerger.