Add `is_auto_mergeable?` & `reasons_not_to_merge` to PolicyManager
In the next commit, we'll move these checks out of PullRequest and
into PolicyManager.

PullRequest should know how to approve and merge PRs, and should
know how to check CI passes, and that the right number of commits
and file changes are a part of the PR. It no longer makes sense
for PullRequest to know about what kinds of dependency update are
allowed to be auto-merged. That should live in PolicyManager.