Add `dependency_policy` method
This method allows us to easily determine whether a given
dependency is auto-mergeable in theory (on the assumption that
the PullRequest checks all pass too). We take into account:

1) Whether or not the dependency is 'external'
2) What the default policy / 'behaviours' are
3) Whether there are any overrides provided

This is really the meat of the logic for RFC-167. We want to be
able to set a series of defaults (or rely on the defaults included
by govuk-dependabot-merger), but also be able to specify overrides
on a per-dependency basis. It all comes together in this method.

We'll make use of this method in the next commit.