Add 'change_allowed?' method
This uses `dependency_policy` under the hood. It is a convenience
method that returns `true`/`false` depending on whether the given
dependency and semver type are considered auto-mergeable.

In the next commit, we'll see the payoff of these two new methods,
which will allow us to delete a lot of redundant methods from
PolicyManager.