Fix `allowed_semver_bumps` config option
These were being treated as arrays of strings, rather than arrays
of symbols, meaning the `change_allowed?` method of PolicyManager
would always return `false` even if the dependency update is of
the correct semver 'size'.

We now cast the config option to an array of symbols at the point
of use, which fixes the issue.