Allow changes to package.json
There are plenty of harmless Dependabot PRs which are currently
not auto-merged because they make changes to the `package.json`
file, which we have not yet configured to allow.
Example PR:
https://github.com/alphagov/travel-advice-publisher/pull/2018

This commit adds that configuration. Similar to the work added
in #84 (which allowed edits to the `Gemfile`), this config tweak
should be safe because:

- CI should continue to catch any breaking change (and prevent
  auto merge)
- Developers can explicitly opt out sensitive repos from auto
  merge, using the govuk_dependabot_merger.yml config file).
- Additionally, devs can configure their dependabot.yml to
  [specify a versioning-strategy](https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuration-options-for-the-dependabot.yml-file#versioning-strategy)
  if they wish Dependabot to only raise PRs that bump the lockfile
  (`lockfile-only` strategy) and ignore any updates that would
  require edits to `package.json`.

This should remove a large proportion of the remaining toil
associated with Dependabot updates. Fixes #85.